requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
customtkinter>=5.2.0
python-slugify>=8.0.0
